% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix_from_hic.R
\name{contact_matrix_from_hic}
\alias{contact_matrix_from_hic}
\title{Extract the contact matrix from a .hic file.}
\usage{
contact_matrix_from_hic(chrn, start, end, reso, fhic, hicnorm)
}
\arguments{
\item{chrn}{The chromosome location.}

\item{start}{The start position of selected contact region.}

\item{end}{The end position of selected contact region.}

\item{reso}{The resolution of bins.}

\item{fhic}{The path of .hic file.}

\item{hicnorm}{Normalization to apply. Must be one of NONE/VC/VC_SQRT/KR.
VC is vanilla coverage, VC_SQRT is square root of vanilla coverage, and KR
is Knight-Ruiz or Balanced normalization.}
}
\value{
The extracted matrix.
}
\description{
Extract the contact matrix from a .hic file.
}
