% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffDomain_Spectrum.R
\name{DiffDomain_Spectrum}
\alias{DiffDomain_Spectrum}
\title{The goodness-of-fit test for the reorganized TADs detection.}
\usage{
DiffDomain_Spectrum(x, y = "semicircle", N = 10000)
}
\arguments{
\item{x}{vector to be test.}

\item{y}{default semicircle.}

\item{N}{sampled numbers to be used to estimate p-value, default 10000.}
}
\value{
a list, include the statistic to estimate p-value, the obtained
p-value, represent as p.value.
}
\description{
The goodness-of-fit test for the reorganized TADs detection.
}
\examples{
x <- runif(100, min = -2, max = 2)
res <- DiffDomain_Spectrum(x)
p_value <- res$p.value
}
