% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffDomain_Spectrum_parallel.R
\name{DiffDomain_Spectrum_parallel}
\alias{DiffDomain_Spectrum_parallel}
\title{Parallel codes for TADs detection.}
\usage{
DiffDomain_Spectrum_parallel(
  tadlist_path,
  fhic0,
  fhic1,
  min_nbin,
  hicnorm,
  reso,
  prop,
  core_num_prop = 0.5,
  seed = 123,
  save_path = "output.txt"
)
}
\arguments{
\item{tadlist_path}{the TADs need to be detected.}

\item{fhic0}{the path of .hic file for condition 1.}

\item{fhic1}{the path of .hic file for condition 2.}

\item{min_nbin}{the minimum number of bins for Hi-C contact matrix,
default 8.}

\item{hicnorm}{Normalization to apply. Must be one of NONE/VC/VC_SQRT/KR.
VC is vanilla coverage, VC_SQRT is square root of vanilla coverage, and KR
is Knight-Ruiz or Balanced normalization.}

\item{reso}{the resolution of bins.}

\item{prop}{a threshold value that represent the proportion of zeros in the
contact matrix, if the proportion is more than this threshold, then the
corresponding rows and columns are removed.}

\item{core_num_prop}{the proportion of core used in parallel calculation.}

\item{seed}{random seed used in the function of future_map().}

\item{save_path}{the save path to save the detection information results.}
}
\value{
all detection results are saved in the save_path file.
}
\description{
Parallel codes for TADs detection.
}
