% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyTADs_HiC.R
\name{identifyTADs_HiC}
\alias{identifyTADs_HiC}
\title{Detect the reorganized TADs between different samples.}
\usage{
identifyTADs_HiC(x, tadlist, fhic0, fhic1, min_nbin, hicnorm, reso, prop)
}
\arguments{
\item{x}{the index of detected TADs in tadlist.}

\item{tadlist}{the TADs need to be detected.}

\item{fhic0}{the path of .hic file for condition 1.}

\item{fhic1}{the path of .hic file for condition 2.}

\item{min_nbin}{the minimum number of bins for Hi-C contact matrix,
default 8.}

\item{hicnorm}{Normalization to apply. Must be one of NONE/VC/VC_SQRT/KR.
VC is vanilla coverage, VC_SQRT is square root of vanilla coverage, and KR
is Knight-Ruiz or Balanced normalization.}

\item{reso}{the resolution of bins.}

\item{prop}{a threshold value that represent the proportion of zeros in the
contact matrix, if the proportion is more than this threshold, then the
corresponding rows and columns are removed.}
}
\value{
a dataframe about the TAD, including the chromosome location, the
start and end of TAD, the region of TAD, and the statistics of calculate
p-value, the p-value of the TAD, and the size of TAD matrix.
}
\description{
Detect the reorganized TADs between different samples.
}
