% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makewindow.R
\name{makewindow}
\alias{makewindow}
\title{Calculate the positions of window in a region.}
\usage{
makewindow(start, end, reso)
}
\arguments{
\item{start}{The start position of a region.}

\item{end}{The end position of a region.}

\item{reso}{The resolution of bin.}
}
\value{
The positions of window.
}
\description{
Calculate the positions of window in a region.
}
\examples{
wins <- makewindow(1, 100000, 10000)
wins
}
