% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normDiffbyMeanSD.R
\name{normDiffbyMeanSD}
\alias{normDiffbyMeanSD}
\title{Normalize the difference matrix.}
\usage{
normDiffbyMeanSD(Diffmat)
}
\arguments{
\item{Diffmat}{A symmetric matrix need to be normalized.}
}
\value{
A normalized symmetric matrix
}
\description{
Normalize the difference matrix.
}
\examples{
A <- matrix(runif(10^2, min = 0, max = 1), nrow = 10)
Diffmat <- (A + t(A)) / 2
mtx <- normDiffbyMeanSD(Diffmat)
mtx
}
