% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvs.R
\name{rvs}
\alias{rvs}
\title{Generate random variables that follow the semi-circle law.}
\usage{
rvs(size, beta = 1, center = 0, sigma = 1)
}
\arguments{
\item{size}{The number of generated random variables.}

\item{beta}{(int, default=1): descriptive integer of the Gaussian ensemble
type. For GOE beta=1, for GUE beta=2, for GSE beta=4.}

\item{center}{(float, default=0.0): center of the distribution.
Since the distribution has the shape of a semicircle, the center corresponds
to its center.}

\item{sigma}{(float, default=1.0): scale of the distribution. This value also
corresponds to the standard deviation of the random entries of the sampled
matrix.}
}
\value{
The generated random variable.
}
\description{
Generate random variables that follow the semi-circle law.
}
\examples{
x <- rvs(10)
x
}
