% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdAdjust.R
\name{sdAdjust}
\alias{sdAdjust}
\title{Calculate the adjust standard deviation.}
\usage{
sdAdjust(numberin)
}
\arguments{
\item{numberin}{A vector to calculate the adjust standard deviation.}
}
\value{
The adjusted standard deviation.
}
\description{
Calculate the adjust standard deviation.
}
\examples{
x <- rnorm(20)
sdAdjust(x)
}
