% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicircle_cdf.R
\name{semicircle_cdf}
\alias{semicircle_cdf}
\title{The distribution function of the semicircle law.}
\usage{
semicircle_cdf(x)
}
\arguments{
\item{x}{vector of quantiles.}
}
\value{
A probability value.
}
\description{
The distribution function of the semicircle law.
}
\examples{
semicircle_cdf(0)
}
